+++ nata-monkey/src/main/java/com/nata/action/MenuAction.java
@@ -17,6 +17,7 @@ public class MenuAction extends Action{
     @Override
     public void fire() {
         device.sendKeyEvent(AndroidKeyCode.MENU);
+        count++;
     }
 
     @Override